+++ lib/src/view/login_screen.dart
@@ -3,7 +3,12 @@ import 'package:get/get.dart';
 import 'package:induk_club_promotion_app_project/src/controllers/login_controller.dart';
 import 'package:induk_club_promotion_app_project/src/responsible_layout.dart';
 import 'package:induk_club_promotion_app_project/src/widget/account_box.dart';
+import 'package:induk_club_promotion_app_project/src/widget/apple_login_button.dart';
+import 'package:induk_club_promotion_app_project/src/widget/google_login_button.dart';
+import 'package:induk_club_promotion_app_project/src/widget/kakao_login_button.dart';
 import 'package:induk_club_promotion_app_project/src/widget/login_box.dart';
+import 'package:induk_club_promotion_app_project/src/widget/login_text_field.dart';
+import 'package:induk_club_promotion_app_project/src/widget/sign_button.dart';
 import 'package:induk_club_promotion_app_project/src/widget/terms_check_box.dart';
 import 'package:induk_club_promotion_app_project/src/widget/verification_box.dart';
 
@@ -15,6 +20,7 @@ class LoginScreen extends StatefulWidget {
 }
 
 class _LoginScreenState extends State<LoginScreen> {
+  final controller = Get.find<LoginController>();
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -38,7 +44,7 @@ class _LoginScreenState extends State<LoginScreen> {
                   height: 50,
                 ),
                 Container(
-                  width: 500,
+                  width: 350,
                   decoration: (ResponsibleLayout.isMobile(context))
                       ? null
                       : BoxDecoration(
@@ -52,12 +58,20 @@ class _LoginScreenState extends State<LoginScreen> {
                           borderRadius: BorderRadius.circular(12.0),
                         ),
                   child: Center(
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        _logo(),
-                        _body(),
-                      ],
+                    child: Padding(
+                      padding:
+                          EdgeInsets.symmetric(vertical: 30, horizontal: 50),
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: [
+                          _logo(),
+                          _textFields(),
+                          _button(),
+                          _options(),
+                          _divider(),
+                          _socialSignUpBtn(),
+                        ],
+                      ),
                     ),
                   ),
                 ),
@@ -70,10 +84,10 @@ class _LoginScreenState extends State<LoginScreen> {
   }
 
   Widget _logo() => Padding(
-        padding: const EdgeInsets.all(20.0),
+        padding: const EdgeInsets.all(10.0),
         child: Container(
-          width: 200,
-          height: 100,
+          width: 100,
+          height: 50,
           alignment: Alignment.center,
           decoration: BoxDecoration(
               border: Border.all(width: 2.0, color: Colors.black)),
@@ -81,20 +95,117 @@ class _LoginScreenState extends State<LoginScreen> {
         ),
       );
 
-  Widget _body() => Padding(
-        padding: (ResponsibleLayout.isMobile(context))
-            ? const EdgeInsets.symmetric(vertical: 20.0, horizontal: 50.0)
-            : const EdgeInsets.symmetric(vertical: 20.0, horizontal: 100.0),
-        child: GetX<LoginController>(builder: (controller) {
-          return IndexedStack(
-            index: controller.pageIndex,
-            children: const [
-              LoginBox(),
-              TermsCheckBox(),
-              AccountBox(),
-              VerificaionBox()
-            ],
-          );
-        }),
+  Widget _textFields() => Padding(
+        padding: const EdgeInsets.symmetric(vertical: 10.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(4.0),
+              child: Text(
+                '이메일',
+                style: Get.theme.textTheme.bodySmall,
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.symmetric(vertical: 4.0),
+              child: LoginTextField(controller: controller.emailController),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(4.0),
+              child: Text(
+                '비밀번호',
+                style: Get.theme.textTheme.bodySmall,
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.symmetric(vertical: 4.0),
+              child: LoginTextField(
+                controller: controller.passwordController,
+                obscureText: true,
+              ),
+            )
+          ],
+        ),
+      );
+
+  Widget _button() => const Padding(
+        padding: EdgeInsets.symmetric(vertical: 10.0),
+        child: SignButton(
+            width: double.infinity,
+            height: 40,
+            child: Text(
+              '로그인',
+              style: TextStyle(color: Colors.white, fontSize: 15),
+            )),
+      );
+
+  Widget _options() {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        GestureDetector(
+            onTap: controller.moveToFindAccount,
+            child: Text(
+              '아이디 찾기',
+              style: Get.theme.textTheme.bodySmall,
+            )),
+        GestureDetector(
+            onTap: controller.moveToFindPassword,
+            child: Text(
+              '비밀번호 찾기',
+              style: Get.theme.textTheme.bodySmall,
+            )),
+        GestureDetector(
+            onTap: controller.moveToTerm,
+            child: Text(
+              '회원가입',
+              style: Get.theme.textTheme.bodySmall,
+            )),
+      ],
+    );
+  }
+
+  Widget _socialSignUpBtn() => Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: GestureDetector(child: const GoogleLoginButton()),
+          ),
+          const Padding(
+            padding: EdgeInsets.all(8.0),
+            child: AppleLoginButton(),
+          ),
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: GestureDetector(
+                onTap: controller.signIn, child: const KakaoLoginButton()),
+          ),
+        ],
+      );
+
+  Widget _divider() => Padding(
+        padding: const EdgeInsets.symmetric(vertical: 10),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
+          children: [
+            Container(
+              width: 100,
+              height: 1.0,
+              decoration: const BoxDecoration(color: Colors.black),
+            ),
+            const Text(
+              "Start with",
+              style: TextStyle(fontSize: 10),
+            ),
+            Container(
+              width: 100,
+              height: 1.0,
+              decoration: const BoxDecoration(color: Colors.black),
+            ),
+          ],
+        ),
       );
 }
